+++ disropt/__init__.py
@@ -1,2 +1,2 @@
 name = "disropt"
-__version__ = "0.1.6"
+__version__ = "0.1.7"