+++ disropt/problems/__init__.py
@@ -3,4 +3,3 @@ from .constraint_coupled_problem import ConstraintCoupledProblem
 from .linear_problem import LinearProblem
 from .quadratic_problem import QuadraticProblem
 from .projection_problem import ProjectionProblem
-from .abstract_opt_problem import AbstractOptimizationProblem